chore: move elevation out of ryot
