+++ crates/ryot/src/bevy_ryot/game/mod.rs
@@ -1,18 +1,14 @@
-use self::elevation::{apply_elevation, Elevation};
 use crate::prelude::*;
 use bevy::prelude::*;
 use ryot_tiled::prelude::*;
 
-pub mod elevation;
 pub mod tile_flags;
 
 pub struct GamePlugin;
 
 impl Plugin for GamePlugin {
     fn build(&self, app: &mut App) {
-        app.add_event::<LoadObjects>()
-            .add_systems(Update, apply_elevation)
-            .add_systems(Last, track_position_changes);
+        app.add_event::<LoadObjects>().add_plugins(ElevationPlugin);
     }
 }
