+++ crates/ryot_content/src/game/mod.rs
@@ -1,1 +1,3 @@
+#[cfg(feature = "bevy")]
+pub mod elevation;
 pub mod game_object;
